\name{LCA4-package}
\alias{LCA4-package}
\alias{LCA4}
\docType{package}
\title{Simulate the LCA}
\description{Simulates the Leaky, Competing Accumulator model (Usher & McClelland, 2001). Contains two functions: LCA() and testLCAsim().}
\details{
  \tabular{ll}{
    Package: \tab LCA4\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2017-08-07\cr
    License: \tab Unlimited\cr
  }
}
\author{
S Miletic
}
